Don't give up when missing subs for one language
This patch fixes an error where the agent would give up and return an early
verdict when it couldn't find any subtitles for a language in the user's
preferred language list - or more specifically, when a result is returned, but
is empty. The expected behavior here should be that it instead exhausts the
list before giving up, in an effort to provide at least one set of subs.